LayoutTests/http/tests/xmlhttprequest/navigation-should-abort-expected.txt

CONSOLE MESSAGE: line 11: PASS: Expected 'error', got 'error'.
If this text shows up, you've successfully navigated to 'navigation-target.html'.
